feat(startup): add operator builder and refactor the startup code.
BREAKING CHANGE: the default way the operator was started was removed.
Please change to the new "generic host" way of starting
the operator. In the Program.cs file, the operator wrapper is still
needed for the additional operator commands.

BREAKING CHANGE: Complete overhaul of the KubernetesOperator class
The migration path will be to create a new asp.net project
and then add the kubernetes operator back again.

BREAKING CHANGE: Removed the kubernetes test operator.
Instead, a web application factory (called `KubernetesOperatorFactory`)
was added to be used by the tests to startup the server and
configure the test application further.

Signed-off-by: Christoph Bühler <christoph@smartive.ch>