feat(startup): add operator builder and refactor the startup code. (#47)
BREAKING CHANGE: the default way the operator was started was removed.
Please change to the new "generic host" way of starting
the operator. In the Program.cs file, the operator wrapper is still
needed for the additional operator commands.

BREAKING CHANGE: Complete overhaul of the KubernetesOperator class
The migration path will be to create a new asp.net project
and then add the kubernetes operator back again.

BREAKING CHANGE: Removed the Kubernetes Operator class
and inserted generic host extensions for the CreateHostBuilder
method of asp.netcore.

BREAKING CHANGE: Removed the kubernetes test operator.
Instead, a web application factory (called KubernetesOperatorFactory)
was added to be used by the tests to startup the server and
configure the test application further.

BREAKING CHANGE: Removed the AddResourceFinalizer extension
method. This part is now found in the resource controller. There
the entity types are clear and a finalizer can be attached
with the client.