feat(leader election): support leader election via V1Lease.
This closes #22. The leader election utilizes V1Lease objects
in the namespace that the operator is running in. If no object
exists, the instance is elected as leader. Later, the config
determines the interval in which the check occurs. When a leader
does not update the "RenewTime", another instance can take leadership.