Tidy up a couple of compiler warnings.
The disabled nullable warning in "ResourceCache.test.cs" is there
because while not a nullable type, the fact that it comes from
deserialized JSON means that null does sometimes come back.

I didn't want to muck with changing the type of TStatus because of
inheritance and possibility of unintended side-effects.