refactor(controller): Refactoring to scoped controllers.
This closes #138.

BREAKING CHANGE: Refactoring of the whole controller instantation.
This removes the `ResourceControllerBase` class. Please migrate to
the new interface `IResourceController`. The basic idea is to provide
scoped access like asp.net api controller. ALL non abstract classes
that implement that interface are registered as controllers on startup.
Whenever an event for a resource happens, a dependency injection
scope is created and the controller is called with the specific action.
There is no need for a base class anymore.