refactor(finalizer): Use scoped finalizer.
BREAKING CHANGE: Refactoring of the finalizer logic. Same as with
the controller, the finalizer are now instantiated via dependency injection
in a scope. As soon as a finalization request happens, the finalizer
manager will search for the adequate finalizer and call it.
The new way of registering finalizer is via the `IFinalizerManager`
to attach a finalizer to an entity.