refactor: Dependency Injection Registration and OperatorBuilder
This allows one controller to support more than one entity type and/or
one entity-type to be managed by more than one controller.
It is not recommended to implement the latter case, as the watchers
will pass the same object version to all controllers
and there will be race conditions as all controllers are called
simultaneously.

All the `...Type` classes and their generic variants are removed in favor of
adding registrations in `IComponentRegistrar`

Additionally, any special logic for instantiating these registrations should be
encapsulated in separate classes such as with `IControllerInstanceBuilder` in
order to avoid violating SRP as much as possible (introducing that much
reflection can really distract from what a class is really meant to handle).

BREAKING CHANGE: `EntityRbacAttribute` and `GenericRbacAttribute`
are now only found when added to Kubernetes "relevant" classes.
Instead of scanning all assemblies for those attributes, only
controllers, finalizers and webhooks may be decorated and considered
when generating the RBAC rules in the generator. To migrate,
add any of the attributes above to a controller, finalizer or webhook
instead of anywhere else in the code.