Make changes to support de-duplication of resource Events
When an event for a given resource is added to the queue, all enqueued events
(those currently waiting for their "Delay" to elapse) for that resource are
unsubscribed from and are thus ignored.

For example, if I were to queue a "NotModified" event for a resource with
a delay of 10 seconds, then 5 seconds later queue a "Deleted" event, the
"Deleted" event would process immediately, and the "NotModified" event would
be ignored and not processed.

This has no effect on events which are currently in the "UpdateResourceData"
or "HandleEvent" steps, as those are already past the de-duplication step.

Signed-off-by: Erin Allison <erin@eallison.us>