feat: Add options for 'generator installer', remove unused -o/-f options from 'generator' base command help, and change some wording  (#596)
Made a few changes to the 'generator installer' command to make it
easier to script with. The command now has `--image-name` and
`--image-tag` options so users can specify these values when the
kustomization YAML is generated, instead of manually editing the file
after the fact. Some of the other options for this command were not
utilizing the McMaster CLI default value feature, so those values have
been added in-line to make it clearer to users what the defaults are if
they do not provide them. I also reworded some of the option
descriptions to be more unambiguous about what is expected to be in
those folders.

Additionally, the `Generator` class no longer inherits from
`GeneratorBase`, and `GeneratorBase` has been renamed to `OutputBase`.
This is because the `--out` and `--format` options were shown under the
`generator` command, but it seems they didn't actually do anything for
_that_ command. Instead, it seems those options are only used by the
sub-commands themselves (`docker`, `crds`, `rbac`, etc) when the
YAML/JSON is generated. So now, only the `crd`, `rbac`, `docker`,
`installer`, and `operator` commands inherit from `OutputBase`. This
will only be a breaking change for users who are unwittingly using the
options on the `generator` command that go unused by the SDK.

If that shouldn't be a breaking change, some dummy options can be
created and hidden under the `generator` command to prevent CLI parsing
errors. However, this will result in unexpected behavior for new users
if they accidentally place their -o/-f options on the wrong command
(resulting in the contents being output to the console), which is
somewhat similar to how it behaved before this PR.

---------

Co-authored-by: Christoph Bühler <buehler@users.noreply.github.com>