refactor: Split Operator in several packages (#605)
BREAKING CHANGE: This removes several features from the
7.x version. Required / missing features will
be added to the v8 again. As for the first
pre release, only the simple watcher and
reconciliation loop is present. No finalizer, events, leadership,
whatsoever is present in the library. The initial pre.0
release is meant to be the base for further implementations.
V8 will - hopefully - contain all required features again.

BREAKING CHANGE: The operator does not have commands
and command line executions packaged into the library.
A new dotnet tool is required for CRD generation and
other generators.

BREAKING CHANGE: All abstractions are extracted into
a KubeOps.Abstractions package.

BREAKING CHANGE: The operator is not required to have
ASP.net as default, it works with a normal console
host application as well.

BREAKING CHANGE: `master` is renamed to `main` and
maintenance and release branches are added.