refactor: CLI Argument parsing (#615)
This adds "System.CommandLine" as default parser
and spectre.console for beautiful output.

BREAKING CHANGE: generator commands
may define a csproj or sln file on where the
entities or other elements are located.
If no file is provided, the current directory
is searched and the command fails if none is found.