feat(web-operator): add conversion webhooks (#639)
This closes #137. This adds the preview version of conversion webhooks.
The API is not stable yet and changes may happen in the future. They
require the preview features flag to be set.

BREAKING CHANGE: The CLI (which broke anyway...) does now
generate everyting together and determines based on the
project if there are webhooks or not. This results in one
config folder instead of separate folders for CRD, RBAC, etc.
With this, conversion webhooks can inject their configuration
into the CRDs. Otherwise, this would be a cumbersome process.
If the need for single topic generation arises again (for example only CRDs),
then these generators can be added again.