refactor(operator): rewrite watchers
The synchronous event-based implementations were replaced by two
separate, now asynchronous IHostedService services. One is responsible
for watching the resources while the other is watching the queue for
manual requeues.

The queue for requeueing entities is now asynchronous as well and
instead of creating multiple timers, it leverages Task.Delay and the
CancellationTokenSource.

Also, finalizer registrations are taking advantage of the ability to store
keyed services in the DI container.