feat: `CancellationToken` support and a lot of async improvements (#725)
This PR adds the ability to cancel (almost) all asynchronous operations
with the built-in `CancellationToken`. Due to the nature of interfaces,
almost all of those changes are breaking the existing API.

It also adds an `WatchAsync` method to the `IKubernetesClient`, which is
making use of the newer asynchronous enumerables.

BREAKING CHANGE: Some methods do now feature the cancellation
token which changed the method signature.