feat(core): Add certificates to Operator.Web (#756)
This adds a new CertificateGenerator to Operator.Web, and it reworks
a handful of the classes in Operator.Web to support using certificates
in code. I more or less re-implemented the BouncyCastle solution from
the CLI in Operator.Web using the built-in .NET classes, and I went to
some length to ensure the generated certificates are virtually identical
to those produced by BouncyCastle.