fix: Leader election failure to restart (#783)
This fixes the following behavioral issues noted when
testing a leader-aware operator with transient network issues:
- In `LeaderElectionBackgroundService`, if
`elector.RunUntilLeadershipLostAsync()` throws, the exception is not
observed in the library and no further attempts to become the leader
occur. The library now logs any unexpected exceptions and tries to
become the leader again.
- A leader could not stop and then subsequently start being a leader
once more due to cancellation token sources not being recreated. The
library now disposes and recreates the cancellation token sources as
required.
- `LeaderAwareResourceWatcher<TEntity>.StoppedLeading` would erroneously
pass a cancelled cancellation token to `ResourceWatcher<TEntity>`. The
library now passes the `IHostApplicationLifetime.ApplicationStopped`
token to the `ResourceWatcher<TEntity>` - we can assume that
`ApplicationStopped` is a good indication that the stop should no longer
be graceful.