fix: When resource watch terminates and re-connects the currentVersion should be reset.
This is because, if the operator has not been functioning for a while the previous currentVersion will be too old and the WatchAsync will return an exception k8s.KubernetesException: too old resource version: xxxxx (xxxx). By moving the current version variable inside of the while loop, the current version will start from the current resource version.

This is desired behavour because, having an operator pod that looks like its working, when actually its not reconciling anything cant be right.