feat: Added option to provide an accessable docker image through a CLI option (#864)
The option to provide an accessible CLI image through `--docker-image`
and its tag through `--docker-tag` has been added to the generate
command. This just makes things a bit nicer generate instead of having
to do a second patch over the top of the generated manifest when, for
example, generating a release for an operator we would no longer need to
do a string replacement to or patch to correct the docker image.

closes #863