fix: Removing the nested Pair (#872)
This pull request refactors the handling of common labels in the
`KustomizationCommonLabels` class and its associated
`KustomizationCommonLabelsPair` class to simplify the data structure and
improve usability. The most important changes include replacing the
`KustomizationCommonLabelsPair` class with a direct inheritance from
`Dictionary<string, string>` and updating the
`KustomizationCommonLabels` class to reflect this change.

### Refactoring of common labels handling:

*
[`src/KubeOps.Abstractions/Kustomize/KustomizationCommonLabelsPair.cs`](diffhunk://#diff-500791e44853948e64cc50e1fbd37cb5ba56cef839636e30ab119a7ea54ba332L3-R3):
Simplified the `KustomizationCommonLabelsPair` class by making it
directly inherit from `Dictionary<string, string>`, removing its
properties and methods.

*
[`src/KubeOps.Abstractions/Kustomize/KustomizationCommonLabels.cs`](diffhunk://#diff-9ff89020ff654e0389952b6f3075c61c281221e538f32dd9854c9586e5c1d469L10-R24):
Updated the `Pairs` property in `KustomizationCommonLabels` to use the
new `KustomizationCommonLabelsPair` structure. Replaced the
initialization logic to add key-value pairs directly to the dictionary.
Removed the `IncludeSelectors` property from
`KustomizationCommonLabelsPair` and reintroduced it as a property in
`KustomizationCommonLabels` with a default value of `true`.

This is a fix for
https://github.com/buehler/dotnet-operator-sdk/issues/868