feat: support inheritance for controllers and finalizers (#873)
@buehler maybe you can double check this PR if you will find some time.

we came accross the issue, that any controller and finalizer need to
implement `IEntityController<TEntity>` or corresponding
`IEntityFinalizer<TEntity>` to generate the appropriate entries to
regiter them in the ServiceCollection. Having as base class like
EntityControllerBase<TEntity> / EntityFinalizerBase<TEntity> which
implements `IEntity...` wouldn't work.

As a workaround I just added a redundant `IEntity...` on the inherited
class as well as on the base class. Not the sweetest solution but a
working one :)

As far as I could see this was due to the pure textual syntax analysis
which just checked for the baselist-types. I've changed this to walk the
semantic syntax model and support inheritance upon multi-levels.

thanks in advance.

---------

Co-authored-by: Christoph Bühler <buehler@users.noreply.github.com>